Colocate guest socket binding type
Keep the broker-authorized binding value alongside the reservation and platform socket APIs.

Copilot-Session: b5a1a347-37a8-4246-8bbc-306590921475
