+++ litebox_broker_core/src/socket.rs
@@ -4,9 +4,8 @@
 //! Broker authority for socket policy, guest-visible endpoints, and platform
 //! lifecycle.
 
-mod network;
-
 use alloc::{sync::Arc, vec::Vec};
+use core::fmt;
 use core::net::{Ipv4Addr, SocketAddrV4};
 use core::sync::atomic::{AtomicUsize, Ordering};
 
@@ -24,8 +23,6 @@ use spin::{Mutex, Once};
 use crate::readiness::{ReadinessRegistration, ReadinessSink};
 use crate::session::{ObjectEntry, ObjectRights};
 use crate::{BrokerError, BrokerSession, Result, SessionId};
-
-pub use network::GuestSocketBinding;
 
 const DEFAULT_TCP_LISTEN_ADDRESS: SocketAddrV4 = SocketAddrV4::new(Ipv4Addr::LOCALHOST, 0);
 const DEFAULT_TCP_LOCAL_ADDRESS: SocketAddrV4 = SocketAddrV4::new(Ipv4Addr::LOCALHOST, 0);
@@ -227,6 +224,76 @@ impl Drop for GuestBindingReservation {
         debug_assert!(removed, "live guest binding reservation must be registered");
     }
 }
+
+/// Broker-authorized guest binding passed to a platform provider.
+#[derive(Clone)]
+pub struct GuestSocketBinding {
+    requested: SocketAddrV4,
+    transport: GuestTransport,
+}
+
+impl GuestSocketBinding {
+    fn new(reservation: &GuestBindingReservation) -> Self {
+        Self {
+            requested: reservation.requested_address(),
+            transport: reservation.transport,
+        }
+    }
+
+    /// Returns the broker-reserved guest-visible binding.
+    #[must_use]
+    pub const fn requested(&self) -> SocketAddrV4 {
+        self.requested
+    }
+
+    /// Returns whether the original binding used the wildcard address.
+    #[must_use]
+    pub const fn is_wildcard(&self) -> bool {
+        self.requested.ip().is_unspecified()
+    }
+
+    /// Checks that this value represents a supported guest binding.
+    #[must_use]
+    pub fn is_valid(&self) -> bool {
+        self.requested.port() != 0 && guest_binding_address_is_valid(self.requested)
+    }
+
+    /// Returns whether this binding belongs to the TCP guest namespace.
+    #[must_use]
+    pub fn is_tcp(&self) -> bool {
+        self.transport == GuestTransport::Tcp
+    }
+
+    /// Returns whether this binding covers one concrete guest loopback address.
+    #[must_use]
+    pub fn covers(&self, address: SocketAddrV4) -> bool {
+        self.is_valid()
+            && address.port() != 0
+            && address.ip().is_loopback()
+            && if self.is_wildcard() {
+                address.port() == self.requested.port()
+            } else {
+                address == self.requested
+            }
+    }
+}
+
+impl fmt::Debug for GuestSocketBinding {
+    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
+        formatter
+            .debug_struct("GuestSocketBinding")
+            .field("requested", &self.requested)
+            .finish_non_exhaustive()
+    }
+}
+
+impl PartialEq for GuestSocketBinding {
+    fn eq(&self, other: &Self) -> bool {
+        self.requested == other.requested && self.transport == other.transport
+    }
+}
+
+impl Eq for GuestSocketBinding {}
 
 /// Platform socket and endpoint metadata returned by an accept operation.
 pub struct AcceptedPlatformSocket {
@@ -1199,7 +1266,7 @@ fn stream_status(object: &spin::RwLock<ObjectEntry>) -> Result<SocketStatusRespo
     if socket.connection_status == SocketConnectionStatus::Connecting {
         socket.connection_status = platform_status.status;
     }
-    // Another status call may have reached a terminal state while this platform query was in flight.
+    // Return the latest broker state rather than this potentially stale platform snapshot.
     Ok(SocketStatusResponse {
         status: socket.connection_status,
         local_address: socket.local_address,
